Additional changes
Created the about, contact, and gallery pages.
Created a template page for future use when whole projects are added. You will have an image of the project in the gallery or on the index page, and then you can click that image and it will take you to the specific page that shows all of the images and information about that page